requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
-e .
streamlit
langchain-google-genai
langchain-pinecone
huggingface-hub
pinecone-client

